fixed bug, getting the quantity
git-svn-id: https://jade.tilab.com/svn/jade_add-ons/trunk@817 815794fe-97f2-0310-9206-e62f04a60ddc

+++ ascml/src/jade/tools/ascml/gui/panels/AgentInstanceGeneral.java
@@ -577,7 +577,7 @@ public class AgentInstanceGeneral extends AbstractPanel implements ActionListene
 		{
 			agentInstance.setName(textFieldInstanceName.getText());
 			agentInstance.setNamingScheme(textFieldScheme.getText());
-			agentInstance.setQuantity(spinnerQuantity.getValue()+"");
+			agentInstance.setQuantity(((Double)spinnerQuantity.getValue()).intValue()+"");
 			agentInstance.setTypeName(textFieldTypeName.getText());
 			agentInstance.setType((IAgentType)getRepository().getModelIndex().getModel(textFieldTypeName.getText()));
